/pages を AppRouter に移動させた
